requirements.txt
aiohttp
APScheduler
async-timeout
chardet
discord.py
discord.py[voice]
emoji
multidict
pytz
six
tzlocal
websockets
python-aiml
wikipedia
